pyproject.toml
[tool.poetry]
name = "flask-background-tasks-monolith"
version = "0.1.0"
description = "flask background tasks monolith app"
authors = ["Juan Correa jp.correa@uniandes.edu.co","Jhonatan Hernandez je.hernandezr@uniandes.edu.co", "Carlos Salazar ca.salazara@uniandes.edu.co"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.17.0"
bcrypt = "^3.2.0"
black = "^22.1.0"
Flask = "^2.0.2"
Flask-Bcrypt = "^0.7.1"
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.3.1"
flask-marshmallow = "^0.14.0"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
isort = "^5.10.1"
Jinja2 = "^3.0.3"
marshmallow = "^3.14.1"
marshmallow-sqlalchemy = "^0.27.0"
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.31"
celery = "^5.2.3"
ffmpeg-python = "^0.2.0"
psycopg2 = "^2.9.3"
redis = "^4.1.3"
gunicorn = "^20.1.0"
wheel = "^0.37.1"
pynamodb = "^5.2.1"
s3fs = "^2022.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
